fix(util): migrate "WorkingDirectory" on Linux back to XDG_DATA_HOME
Stacked and dependent on #1964.

**Please review the latest commit only**.

Previously we change this directory in #1948, but this may cause we
could not locate `~/.minecraft` as for assets. PR #1965 lets this
directory configurable.

Here this patch requests don't use ~/.xxxx. This might mess up our
users' home directory. Many UNIX programs, like tmux, nvim, fish... have
adopted to the XDG specification.

As for `~/.minecraft`, I personally feel there are no compatibility
issues because now we only use this directory for the assets (currently,
skins), and these contents are recoverable from HMCL automatically.