chore(l10n): revise description of copy code option on terracotta page
Revise to make it clearer for new users.

The relevant text remains in Simplified Chinese characters in
Traditional Chinese, as PCL CE is currently only available
in Simplified Chinese.

Remove the description of PCL CE in English due to the excessive
length of the text and the language situation of PCL CE mentioned
in the previous paragraph.